changes to update to awestruct 0.5.2
