+++ _ext/jdf.rb
@@ -3,209 +3,6 @@ require 'git'
 require 'fileutils'
 require 'rexml/document'
 
-class Artifact
-  attr_reader :name, :coordinates
-
-  def initialize(name, coordinates)
-    @name = name
-    @coordinates = coordinates
-  end
-
-  class Coordinates
-    attr_reader :groupId, :artifactId, :packaging, :version
-    def initialize(groupId, artifactId, packaging = :jar, version = nil)
-      @groupId = groupId
-      @artifactId = artifactId
-      @packaging = packaging.nil? ? :jar : packaging.to_sym
-      @version = version
-    end
-
-    def self.parse(coordinates_str)
-      # The * spreads the array across the arguments of the constructor
-      new(*coordinates_str.split(':'))
-    end
-
-    def to_maven_dep(opts = {})
-      scope = opts[:scope] || :compile
-      include_version = opts[:include_version].nil? ? true : opts[:include_version]
-      buf = "<dependency>\n".
-          concat("  <groupId>#{@groupId}</groupId>\n").
-          concat("  <artifactId>#{@artifactId}</artifactId>\n")
-      buf.concat("  <type>#{@packaging}</type>\n") if !packaging.to_sym.eql? :jar
-      buf.concat("  <version>#{@version}</version>\n") if include_version and !version.nil?
-      buf.concat("  <scope>#{scope}</scope>\n") if !scope.to_sym.eql? :compile
-      buf.concat("</dependency>")
-    end
-
-    def to_url(base_url = 'http://repo1.maven.org/maven2')
-      [base_url, @groupId.gsub('.', '/'), @artifactId, @version,
-          @artifactId + '-' + @version + '.' + @packaging.to_s] * '/'
-    end
-
-    def to_pom_url(base_url = 'http://repo1.maven.org/maven2')
-      [base_url, @groupId.gsub('.', '/'), @artifactId, @version,
-          @artifactId + '-' + @version + '.' + @packaging.to_s].join('/').gsub(/\.jar$/, '.pom')
-    end
-
-    def to_relative
-      [@artifactId, @packaging.to_s, @version] * ':'
-    end
-
-    def to_s
-      [@groupId, @artifactId, @packaging.to_s, @version] * ':'
-    end
-  end
-end
-
-module Awestruct
-  module Handlers
-    class AsciidocHandler
-
-      def rendered_content(context, with_layouts=true)
-        out_file="#{site.tmp_dir}/asciidoc#{relative_source_path.gsub(/\.([^.]+)$/, '.html')}"
-        if !File.exists?(out_file)
-          FileUtils.mkdir_p(File.dirname(out_file))
-          puts "Processing asciidoc #{relative_source_path} -> #{out_file}"
-          cmd="asciidoc -b html5 -a pygments -a icons -a iconsdir='#{site.base_url}/images' -a imagesdir='../' -o '#{out_file}' #{context.page.source_path}"
-          `#{cmd} `
-        end
-        output = File.open(out_file).read
-
-        extract_metadata(context.page, output)
-       
-      end
-
-      # Proofing our call to asciidoc for future versions
-      def execute_shell(command, input = nil)
-        Open3.popen3(command) do |stdin, stdout, stderr, thrd|
-          stdin.puts input unless input.nil?
-          stdin.close
-          SystemCallError.new("Error invoking #{command}, #{stderr.read}", thrd.value.exitstatus)
-          out = stdout.read
-        end
-      rescue Errno::EPIPE
-        ""
-      end
-
-      def extract_metadata(page, output)
-        html = Nokogiri::HTML(output)
-        # Asciidoc renders a load of stuff at the top of the page, which we need to extract bits of (e.g. author, title) but we want to dump it for rendering
-        page.source_title = html.css("h1").first.text
-        guide_content = html.css('div#content').first 
-        guide_content['id'] = 'guide-content'
-        guide_content['class'] = 'asciidoc'
-        # Extract authors
-        author = html.css('span#author').first
-        if author
-          page.source_author = [ author.text ]
-        end
-        page.source_summary = first_x_words(guide_content.css('p').first.text, 25, '').gsub("\n", '')
-
-        # rebuild links
-        guide_content.css('a').each do |a|
-          if a['href'] =~ /^#(\w*)-(\w*)$/
-            a['href'] = "../#{$1}/##{$1}-#{$2}"
-          end
-        end
-
-        guide_content.to_html
-      end
-
-      def first_x_words(str,n=20,finish='&hellip;')
-        str.split(' ')[0,n].inject{|sum,word| sum + ' ' + word} + finish
-      end
-
-    end
-  end
-end
-
-module Awestruct
-  module Handlers
-    class MarkdownHandler
-
-      def rendered_content(context, with_layouts=true)
-        doc = RDiscount.new( delegate.rendered_content( context, with_layouts ) )
-        extract_metadata(context.page, doc.to_html)
-      end
-
-      def extract_metadata(page, output)
-        
-        html = Nokogiri::HTML(output)
-
-        # Strip out html and body
-        guide_content = html.css('body').first
-        guide_content.name = 'div'
-        guide_content['id'] = 'guide-content'
-        guide_content['class'] = 'markdown'
-
-        # Markdown doesn't have an metadata syntax, so all we can do is pray ;-)
-        # Look for a paragraph that contains tags, which we define by convention
-        # Remove if found
-        guide_content.css('p').each do |p|
-          extract(page, p, 'Author', true)
-          extract(page, p, 'Technologies', true)
-          extract(page, p, 'Level')
-          extract(page, p, 'Summary')
-          extract(page, p, 'Prerequisites', true)
-          extract(page, p, 'Target Product', true)
-        end
-        
-        # Strip out title
-        h1 = guide_content.css('h1').first
-        if h1
-          page.source_title = h1.text
-          h1.remove
-        end
-
-        metadata = guide_content.css('p#metadata').remove
-
-        guide_content.css('h2').each do |header_html|
-          link_id = header_html.inner_html.gsub(' ', '_').gsub('&#8217;', '_').gsub(/[\(\)\.!]/, '').gsub(/\?/, '').downcase
-          header_html['id'] = link_id
-        end
-
-        # rebuild links
-        guide_content.css('a').each do |a|
-          if a['href'] =~ /^(.*\/)README.md(.*)$/
-            a['href'] = $1 + $2
-          end
-        end
-        
-        guide_content.to_html
-      end
-
-      def find(p, tag)
-        if p.text
-          r = p.text[/^(#{tag}: )(.+)$/, 2]
-          if r
-            p['id'] = 'metadata'
-            return r 
-          end
-        end
-      end
-
-      def extract(page, p, tag, split=false)
-        if split
-          s = find_split(p, tag)
-        else
-          s = find(p, tag)
-        end
-        if s
-          page.send("source_#{tag.downcase}=".sub(' ','_').to_sym, s)
-        end
-      end
-
-      def find_split(p, tag)
-        s = find(p, tag)
-        if s
-          return s.split(',').sort 
-        end
-      end
-
-    end
-  end
-end
-
 
 module Awestruct::Extensions::Repository::Visitors
   module Clone
