+++ quickstarts/get-started/index.html.haml
@@ -62,7 +62,7 @@ description: Get started using the quickstarts. JBoss Developer Framework shows 
           %td= guide.technologies ? guide.technologies.join(', ') : '' 
           %td= guide.summary ? guide.summary : ''
           %td= guide.level ? guide.level : ''
-          %td= guide.target_product ? guide.target_product.join(', ') : ''
+          %td= guide.target_product ? guide.target_product : ''
           %td= guide.prerequisites ? guide.prerequisites.join(', ') : ''
 
 :javascript
